+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+* `Vector::<T>::splat` where `T: Copy` to easily create vector with the same value for `x` and `y`
+
 
 ## [0.1.1] - 2023-12-03